remove new call query
