+++ apps/backend/src/datasources/mockups/RoleDataSource.ts
@@ -1,10 +1,8 @@
-import { Call } from '../../models/Call';
 import { Instrument } from '../../models/Instrument';
 import { Role } from '../../models/Role';
 import { CreateRoleArgs } from '../../resolvers/mutations/CreateRoleMutation';
 import { UpdateRoleArgs } from '../../resolvers/mutations/UpdateRoleMutation';
 import { RoleDataSource, Tag } from '../RoleDataSource';
-import { dummyCalls } from './CallDataSource';
 import { dummyInstruments } from './InstrumentDataSource';
 
 const dummyRole: Role = {
@@ -54,9 +52,6 @@ export class RoleDataSourceMock implements RoleDataSource {
       permissions: [],
       isRootRole: false,
     } as Role;
-  }
-  async getCallsByRoleId(agent: number): Promise<Call[]> {
-    return dummyCalls;
   }
   async getInstrumentsByRoleId(agent: number): Promise<Instrument[]> {
     return dummyInstruments;
