+++ apps/backend/src/resolvers/queries/CallsQuery.ts
@@ -110,9 +110,4 @@ export class CallsQuery {
   callsOfReviewer(@Ctx() context: ResolverContext) {
     return context.queries.call.getCallsOfReviewer(context.user);
   }
-
-  @Query(() => [Call])
-  callsByRole(@Ctx() context: ResolverContext) {
-    return context.queries.call.getTagCallsByRoleId(context.user);
-  }
 }
