--- apps/frontend/src/graphql/call/getCallsByRole.graphql
@@ -1,6 +1,0 @@
-query getCallsByRole {
-  callsByRole {
-    ...call
-    
-  }
-}
